+++ changelog.md
@@ -1,3 +1,11 @@
 # Changelog
 
-All online tests work after code merging.
+**./test-parts/08-api_requests/sub/common-api.js**
+* Re-wrote 'testPositiveNumberProperty' to use 'commonFunctions'
+* Removed 'commonPaths.testObject' requirement.
+
+---
+
+**./test-parts/08-api_requests/sub/common-health.js**
+* Re-wrote 'testPositiveNumberProperty' to use 'commonFunctions'
+* Removed 'commonPaths.testObject' requirement.